电商转介绍活动注册页面功能

See merge request !18

+++ esp-domain/src/main/java/com/apass/esp/domain/entity/AwardBindRel.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class AwardBindRel {
     private Long id;
 
-    private Byte activityId;
+    private Long activityId;
 
     private Long userId;
 
@@ -29,15 +29,15 @@ public class AwardBindRel {
         this.id = id;
     }
 
-    public Byte getActivityId() {
-        return activityId;
-    }
+    public Long getActivityId() {
+		return activityId;
+	}
 
-    public void setActivityId(Byte activityId) {
-        this.activityId = activityId;
-    }
+	public void setActivityId(Long activityId) {
+		this.activityId = activityId;
+	}
 
-    public Long getUserId() {
+	public Long getUserId() {
         return userId;
     }
